Add scala_grpc_zio conflict resolver strategy (#114)
* Add scala_grpc_zio conflict resolver

* Add scala_grpc_zio conflict resolver

* Code review

* Improve 'name' collection when parsing scala files (#115)

* Improve name parsing

In a bazel package having multiple scala_library rules of the same package, there may be symbols
that are not imported but still valid.  For example, if "A.scala" has package "foo.bar" and defines
class A, and "B.scala" defines class B extends A, "A" does not necessarily require an import if they
are both in the package "foo".  However, the rule "b_scala" needs "a_scala" in its deps.  How to
determine that?  This is accomplished by running the resolve action twice.  The first time the scopes
are populated, and the second time, we can use the "gazelle:resolve_file_symbol_name" directive to
try and autopopulate imports based on completion of symbols in the "Names" file metadata.